Move ScannerCommand class to lib directory
Oclif was detecting this abstract class as a command, which caused it to
show up in help commands. Move the file to the lib directory in order to avoid this issue.

Upper cased file since it contains a class.
