+++ src/commands/scanner/rule/list.ts
@@ -2,7 +2,7 @@ import {flags} from '@salesforce/command';
 import {Messages} from '@salesforce/core';
 import {Controller} from '../../../ioc.config';
 import {Rule} from '../../../types';
-import {ScannerCommand} from '../scannerCommand';
+import {ScannerCommand} from '../../../lib/ScannerCommand';
 
 // Initialize Messages with the current plugin directory
 Messages.importMessagesDirectory(__dirname);
